localnet/manage.sh: diagnostics output
If a subcommand fails for some unforeseen reason, the script just silently terminates.
This is not the most optimal behaviour if a person runs the script for the first time.

With these changes the script may be run as follows:
```
VERBOSE_OUT=/dev/fd/1 /bin/bash -x ./localnet/manage.sh
```